cleanup tracing
