+++ lib/tracing/cds.js
@@ -5,13 +5,27 @@ const APPLOG = cds.log('app')
 const trace = require('./trace')
 const wrap = require('./wrap')
 
-function wrapEventHandler(eventHandler) {
-  const phase = eventHandler.on ? 'on' : eventHandler.before ? 'before' : 'after'
-  eventHandler.handler = wrap(eventHandler.handler, {
+function _wrapHandler(handler) {
+  const phase = handler.on ? 'on' : handler.before ? 'before' : 'after'
+  handler.handler = wrap(handler.handler, {
     loggerName: APPLOG.label,
     phase: phase,
-    event: eventHandler[phase]
+    event: handler[phase]
   })
+}
+
+const _wrapStmt = (stmt, impl, sql) => {
+  for (const fn of ['run', 'get', 'all', 'stream', 'runBatch']) {
+    if (!stmt[fn]) continue
+    const it = stmt[fn]
+    stmt[fn] = wrap(it, {
+      no_locate: true,
+      wrapper: function () {
+        return trace(`${impl} - stmt.${fn} ${sql}`, it, this, arguments, { loggerName: LOG.label })
+      }
+    })
+  }
+  return stmt
 }
 
 module.exports = () => {
@@ -30,6 +44,17 @@ module.exports = () => {
     }
   })
 
+  const { spawn } = cds
+  cds.spawn = wrap(spawn, {
+    wrapper: function () {
+      const handlerFn = typeof arguments[0] === 'function' ? arguments[0] : arguments[1]
+      const wrappedFn = wrap(handlerFn, { event: 'cds.spawn Handler', loggerName: LOG.label })
+      if (typeof arguments[0] === 'function') arguments[0] = wrappedFn
+      else arguments[1] = wrappedFn
+      return trace({ phase: '', event: 'Background Process' }, spawn, this, arguments, { loggerName: LOG.label })
+    }
+  })
+
   cds.on('serving', service => {
     // Do trace event handler either when
     // Logging is enabled for all and it is not explicitly  disabled for this service
@@ -40,7 +65,7 @@ module.exports = () => {
       (service.definition['@cds.tracing'] && APPLOG.level !== 0)
     ) {
       for (const each of ['_error', '_initial', 'on', 'before', 'after']) {
-        service._handlers[each].forEach(wrapEventHandler)
+        service._handlers[each].forEach(_wrapHandler)
       }
     }
 
@@ -53,15 +78,25 @@ module.exports = () => {
     }
   })
 
-  // REVISIT: Does it also work with cds.context.spawn?
-  const { spawn } = cds
-  cds.spawn = wrap(spawn, {
-    wrapper: function () {
-      const handlerFn = typeof arguments[0] === 'function' ? arguments[0] : arguments[1]
-      const wrappedFn = wrap(handlerFn, { event: 'cds.spawn Handler', loggerName: LOG.label })
-      if (typeof arguments[0] === 'function') arguments[0] = wrappedFn
-      else arguments[1] = wrappedFn
-      return trace({ phase: '', event: 'Background Process' }, spawn, this, arguments, { loggerName: LOG.label })
-    }
-  })
+  const impl = cds.env.requires.db?.impl
+  if (impl?.match(/^@cap-js\//)) {
+    const dbService = require(impl)
+    cds.once('served', () => {
+      const { prepare, exec } = dbService.prototype
+      dbService.prototype.prepare = wrap(prepare, {
+        wrapper: function (sql) {
+          const stmt = trace(`${impl} - prepare ${sql}`, prepare, this, arguments, { loggerName: LOG.label })
+          if (stmt instanceof Promise) return stmt.then(stmt => _wrapStmt(stmt, impl, sql))
+          return _wrapStmt(stmt, impl, sql)
+        }
+      })
+      dbService.prototype.exec = wrap(exec, {
+        wrapper: function (sql) {
+          if (!cds.env.requires.telemetry.tracing.tx && sql in { BEGIN: 1, COMMIT: 1, ROLLBACK: 1 })
+            return exec.apply(this, arguments)
+          return trace(`${impl} - exec ${sql}`, exec, this, arguments, { loggerName: LOG.label })
+        }
+      })
+    })
+  }
 }
