+++ lib/tracing/cloud_sdk.js
@@ -4,6 +4,7 @@ const LOG = cds.log('cds')
 const trace = require('./trace')
 const wrap = require('./wrap')
 
+// REVISIT: unverified!
 module.exports = () => {
   try {
     require.resolve('@sap-cloud-sdk/http-client')
