+++ lib/tracing/index.js
@@ -5,7 +5,7 @@ const { trace, SpanKind } = require('@opentelemetry/api')
 const { registerInstrumentations } = require('@opentelemetry/instrumentation')
 const { BatchSpanProcessor, SimpleSpanProcessor } = require('@opentelemetry/sdk-trace-base')
 const { NodeTracerProvider } = require('@opentelemetry/sdk-trace-node')
-const { SemanticAttributes /* , SemanticResourceAttributes */ } = require('@opentelemetry/semantic-conventions')
+const { SemanticAttributes } = require('@opentelemetry/semantic-conventions')
 
 const { _require } = require('../utils')
 
@@ -112,10 +112,8 @@ module.exports = resource => {
   const instrumentations = _getInstrumentations()
   registerInstrumentations({ tracerProvider, instrumentations })
 
-  // REVISIT: why not do if dynatrace is in deps?
   if (process.env.NODE_ENV !== 'production' || !_isDynatraceEnabled()) {
     const exporter = _getExporter()
-    // REVISIT: why this difference?
     if (process.env.NODE_ENV !== 'production') {
       tracerProvider.addSpanProcessor(new SimpleSpanProcessor(exporter))
     } else {
@@ -130,7 +128,6 @@ module.exports = resource => {
    * add CAP instrumentations
    */
   require('./cds')()
-  require('./db')()
   require('./okra')()
   require('./cloud_sdk')()
 }
